--- errors/errors.go
@@ -1,6 +1,0 @@
-package errors
-
-import "errors"
-
-var InvalideEmail = errors.New("invalide email")
-var LostConnection = errors.New("Redis connection lost")
